refactor: split CLI cloudflare helpers
